2007-10-31  Mauricio Vieira  <mauricio.vieira@gmail.com>
	reviewed by: Johan Dahlin  <jdahlin@async.com.br>

	* AUTHORS (Maintainers): Add Mauricio

	* kiwi/ui/gazpacholoader.py:
	* kiwi/ui/widgets/scale.py:
	* tests/test_Scale.py:
	* tests/test_proxy.py:
	* tests/test_utils.py:
	Add support for Proxy scale widget, fixes #3578

+++ kiwi/ui/gazpacholoader.py
@@ -227,7 +227,7 @@ class HScaleDataType(DataTypeAdaptor):
 
 class VScaleDataType(DataTypeAdaptor):
     widget_type = ProxyVScale
-    default = int
+    default = float
 
 class EntryDataType(DataTypeAdaptor):
     widget_type = ProxyEntry
